requirements.txt
aiogram
python-dotenv
pyrogram
tgcrypto
psycopg2-binary